refactor email notification config
