+++ webapp/src/main/resources/static/ts/src/bundles/notification-settings.ts
@@ -1,16 +1,16 @@
 import { AlertService } from "../service/alert-service";
 import { LoadingIndicatorService } from "../service/loading-indicator-service";
-import { NotificationSettingsRenderService } from "../service/notification-settings-render-service";
+import { EmailNotificationSettingsRenderService } from "../service/email-notification-settings-render-service";
 import { NotificationSettingsRestClient } from "../clients/notification-settings-rest-client";
 
 const alertService = new AlertService();
 const loadingIndicatorService = new LoadingIndicatorService();
 const notificationSettingsRestClient = new NotificationSettingsRestClient();
-const notificationRenderService = new NotificationSettingsRenderService(
+const emailNotificationSettingsRenderService = new EmailNotificationSettingsRenderService(
     notificationSettingsRestClient,
     alertService,
     loadingIndicatorService,
 );
 
 const notificationSettings = notificationSettingsRestClient.fetchNotificationSettings();
-notificationRenderService.render(notificationSettings);
+emailNotificationSettingsRenderService.render(notificationSettings);
