+++ webapp/src/main/resources/static/ts/src/model/notification-settings.model.ts
@@ -1,8 +1,21 @@
 export interface NotificationSettings {
+    readonly emailConfig: DefaultNotificationConfig;
+    readonly telegramConfig: TelegramNotificationConfig;
+}
+export interface DefaultNotificationConfig {
     readonly notify: boolean;
     readonly frequencyInWeeks: number;
     readonly notificationAtReleaseDate: boolean;
     readonly notificationAtAnnouncementDate: boolean;
+}
+export interface TelegramNotificationConfig extends DefaultNotificationConfig {
     readonly telegramNotificationsActive: boolean;
     readonly registrationId: number;
+}
+export interface UpdateNotificationSettingsRequest extends DefaultNotificationConfig {
+    readonly channel: NotificationChannel;
+}
+export enum NotificationChannel {
+    EMAIL = "EMAIL",
+    TELEGRAM = "TELEGRAM",
 }
